% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunchtime.r
\name{lunchtime}
\alias{lunchtime}
\title{Lunch Time}
\usage{
lunchtime()
}
\value{
string "It's lunch time". That's all it does. Adding some additional text to trigger GH-actions. Some extra changes.
}
\description{
adding a description to test GH-Action
}
\examples{
lunchtime()

}
